+++ src/Molecules/Button/components/BaseButton/Button.tsx
@@ -88,6 +88,9 @@ export const Button: BaseButtonComponent = React.forwardRef<
     delayLoadingSpinnerAnimation = true,
     icon,
     iconPlacement = 'left',
+    // componentId is sent to trackContext.track() but we don't want to spread it with ...rest
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    componentId,
     ...rest
   } = props;
   const externalIsNotPresent = typeof external === 'undefined';