+++ src/common/PopOver/PopOver.types.ts
@@ -35,6 +35,8 @@ export type Props = {
   handleMouseLeave?: (e: any) => void;
   customBoundary?: HTMLElement | Array<HTMLElement>;
   bottomSheet?: boolean;
+  bottomSheetTitle?: string;
+  onBottomSheetClose?: () => void;
   pointerArrow?: boolean;
   invertedColors?: boolean;
   setPopoverElement?: (r: any) => void;